[enh] Add File based health check module.
This health check module works by comparing the difference of the last
time a file was modified and the current time to a configurable timeout.
If the difference is greater than the timeout, the health check fails.

This implements a _passive_ health check system in which the worker notifies
with an hearbeat that it is still alive.